requirements.txt
biopython>=1.79
pandas>=1.5.3
numpy>=1.24.3
torch>=2.0.1
tqdm>=4.65.0
scipy>=1.10.1
matplotlib>=3.7.1
scikit-learn>=1.2.2
bayesian-optimization>=1.4.1
openpyxl